requirements.txt
# Project HELIX v2.0 - Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI Model SDKs  
openai==1.3.7
anthropic==0.7.8

# Async Support
httpx==0.25.2
aiofiles==23.2.1
aiohttp==3.9.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0

# Logging
structlog==23.2.0